+++ authentication/templates/authentication/reset_success.html
