+++ authentication/templates/authentication/signup_email.html
