+++ core/settings.py
@@ -38,7 +38,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'libertas'
+    'libertas',
+    'authentication'
 ]
 
 MIDDLEWARE = [
