+++ libertas/models.py
@@ -1,26 +1,3 @@
 from django.db import models
-from django.contrib.auth.models import User
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 
 # Create your models here.
-
-
-class Profile(models.Model):
-    user = models.OneToOneField(User, on_delete=models.CASCADE)
-    email_confirmed = models.BooleanField('E-Mail bestätigt', default=False)
-    notes = models.TextField('Interne Notizen', blank=True, default='')
-
-    def __str__(self):
-        return self.user.username
-
-    class Meta:
-        verbose_name = 'Profil'
-        verbose_name_plural = 'Profile'
-
-
-@receiver(post_save, sender=User)
-def update_user_profile(sender, instance, created, **kwargs):
-    if created:
-        Profile.objects.create(user=instance)
-    instance.profile.save()
